# ROX-28941: Use GetByQueryFn in alerts