--- a/central/alert/datastore/datastore.go
+++ b/central/alert/datastore/datastore.go
@@ -30,7 +30,6 @@ type DataStore interface {
 	SearchRawAlerts(ctx context.Context, q *v1.Query, excludeResolved bool) ([]*storage.Alert, error)
 	SearchListAlerts(ctx context.Context, q *v1.Query, excludeResolved bool) ([]*storage.ListAlert, error)
 
-	GetByQuery(ctx context.Context, q *v1.Query) ([]*storage.Alert, error)
 	WalkByQuery(ctx context.Context, q *v1.Query, db func(d *storage.Alert) error) error
 	WalkAll(ctx context.Context, fn func(alert *storage.ListAlert) error) error
 	GetAlert(ctx context.Context, id string) (*storage.Alert, bool, error)