This PR uses a feature created in #15010  to improve alerts service performance and present how to use `GetByQueryFn` in other places.


```
                                          │   old.txt   │              new.txt               │
                                          │   sec/op    │   sec/op     vs base               │
AlertDatabaseOps/searchWithRawListAlert-8   9.708m ± 1%   9.448m ± 1%  -2.68% (p=0.000 n=20)

                                          │   old.txt    │               new.txt               │
                                          │     B/op     │     B/op      vs base               │
AlertDatabaseOps/searchWithRawListAlert-8   8.334Mi ± 0%   8.327Mi ± 0%  -0.09% (p=0.000 n=20)

                                          │   old.txt   │              new.txt               │
                                          │  allocs/op  │  allocs/op   vs base               │
AlertDatabaseOps/searchWithRawListAlert-8   109.2k ± 0%   109.2k ± 0%  +0.00% (p=0.000 n=20)
```

More bench tests are in:
- https://github.com/stackrox/stackrox/pull/14636