@janisz 2025-05-05
```diff
-   grpc_req_duration..............: avg=86.26ms  min=83.74ms med=86.34ms max=90.55ms  p(90)=88.26ms  p(95)=89.34ms 
+  grpc_req_duration..............: avg=75.89ms  min=72.33ms  med=75.5ms   max=81.65ms  p(90)=80.22ms  p(95)=80.89ms 
```